backend/requirements.txt
# Math Routing Agent: Backend Requirements
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
qdrant-client==1.7.3
python-multipart==0.0.6
numpy==1.24.4
scikit-learn==1.3.2
openai==1.3.9
python-dotenv==1.0.0
tavily-python==0.3.3
requests==2.31.0

# LangGraph Agent Framework (Compatible versions)
langchain==0.1.16
langchain-core==0.1.52
langgraph==0.0.55

# DSPy Human Feedback  
dspy-ai==2.4.9

# Additional utilities
pandas==2.0.3
httpx==0.25.2
# Pin huggingface_hub for sentence-transformers compatibility
sentence-transformers==2.2.2
huggingface_hub==0.16.4
weasyprint==60.2
# System Monitoring and Performance
psutil==5.9.6
cachetools==5.3.2

# Translation Support

# Mathematical Processing
sympy==1.12
matplotlib==3.8.2
latex2mathml==3.76.0

# Caching (Redis - optional for production)
redis==5.0.1

# Enhanced Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
structlog==23.2.0
